node can make module.
Module is a set of specific variables or functions, it can be a program or can be a part of different program.
Module gives reusability for javascript code.
module.exports assign variables or functions for different javascript program.
And in different javascript program can use require method. It needs arguments, file path.